refactor(ImageNote): rename `EditorCoreX` to `EditorCore`
