+++ src/islands/note/components/notes/ImageNote/components/EditorCore.tsx
