+++ src/islands/note/components/notes/ImageNote/components/ScanedImageEditor.tsx
@@ -1,6 +1,6 @@
 import { Show, createEffect } from 'solid-js'
 import { Dialog, createDialog } from '../../../utils/Dialog'
-import EditorCore from './EditorCoreX'
+import EditorCore from './EditorCore'
 import type { Sheets } from './Sheet'
 import type { ImageNoteData } from '../types'
 import type { SetStoreFunction } from 'solid-js/store'
